Add mocha for testing javascript code
Added dependencies to bring in mocha js alongside chai and sinon, for providing
assertions and mocking capabilities. JSDom is used to provide a simulated DOM
environment in pure js.

The default setup wires them up, sets the required mocha options and adds a helper
to render markup, parse it through JSDom and return a document object that simulates
a webpage.