Add ResearchSession model
This represents our main resource containing answers to some questions
from which we can generate our consent form / information sheet.

This commit is mostly about modelling validations conditionally in
a "building up our model bit by bit" fashion.

We need to bring our steps into the model for now – it's not ideal but
it's expedient.

Use the custom validator we previously wrote to validate the
methodologies step. At present, the redundant name step is yet to be
removed, which we acknowledge in the spec.

There's a lot of melding of concerns here. Ideally our Steps would be
their own separate resource with their own separate controller.