Fixed some issues with docker compose and docker
Docker compose was mounting the entire dev folder off a local machine
and this caused problems using node modules. Only the files with source
code or tests should be mounted.

The dockerignore file worked ok for linux but not on windows, and caused
docker to not build with docker compose, and copy files it shuldn't with
docker build.

Fixed issue with docker build file not building yarn/npm dependencies

Renamed the web process to rails in docker compose in prep for adding
other containers for webpack watch process and maybe auto linting and
unit testing.