Add rails-controller-testing gem
We don't really want to use this. We'd rather use Cucumber, but we're
keeping that complexity out of this branch for now (partly because
adding Cucumber is already underway in another branch). We expect that
when this branch is shown to work we'll retrofit the Cucumber acceptance
and remove this gem.