Integrate the wizard into the views
Also establish a rudimentary instancing scheme for new research
sessions. We need something better with some design behind it.

We only have a couple of views, one of which is soon to be
decommissioned – age and name. These have been sufficient to show how
the wizard works, largely by switching from form_tag to form_for

The views use the wizard paths now, and the research session is tracked
by keeping its ID in session[:current_research_session_id]. This isn't
ideal for lots of reasons, but it gets us going with the wizard and we
can change that behaviour soon.

There's also a rudimentary _error_summary partial, which just spits out
the error messages on the model in a list. It needs rendering in every
partial presently.