Add partials for each remaining step
This gets us to a blank session preview page at
/research_session/preview with a completed ResearchSession in the
database.

The absence of acceptance tests here is unacceptable corner-cutting,
because this commit highlights the need for us to be acknowledging
existing model values and shows that the _tag form of the helpers is
limiting and verbose.

We need to retrofit Cucumber soonest.