Renamed dynamic to highlight
The use of the word dynamic wasn't clear what it meant. Wanted to make the helper more generic and reusable. Also ties the helper function to the associated css class.

The use of the strong tag helps to signify to screens readers that it is significant (which is why something is highlighted).