Warn user when exiting page if changes made
Some, but not all, browsers warn the user if they try to leave a page
that has changes made to form values unless they submit the form. This
small change polyfills older browsers to also do this.

The reason the event is added and taken away is that if we just try to
return false or something from that event handler it puts that on the
screen. The easiest way to handle things is remove the even handler
altogether when you submit a page as it's no longer needed.