Prefer #humanize/inflections over translations
If we use I18n.translate, translations have to exist in en.yml,
otherwise errors make it into the markup (see below).
String#humanize doesn't have this issue
(but still lets us specify our own translations).

Before, in _progress output (because we didn't have en.time_equipment):
<a href="/questions/time_equipment">
  <span class="translation_missing"
        title="translation missing: en.time_equipment">
        Time Equipment
  </span>
</a>

After, including a #human call in the initializers/inflections.rb:

<a href="/questions/time_equipment">Time / equipment</a>

Because the rest of the progress titles are conventional, (for example,
:methodologies.to_s.humanize => 'Methodologies') we don't need
special cases for anything else and can remove the obligation to
translate in en.yml.