Rename FormHelpers -> FormTagHelper
In reality the methods in this helper correspond to ActionView's own
FormTagHelper in style. Since we're using the singular elsewhere, use it
here too.

Also, move the specs for those methods into form_tag_helper_spec.rb. It
does make the file somewhat large, but it is conventional this way (and
editor tooling to skip between tests and code doesn't break)