Convert radio_group_vertical to form_with
Demonstrate on the first (researcher) page only. Rather than rendering
radio markup ourselves, we delegate it to Rails's
collection_radio_buttons. This has the effect of having both values
and labels filled in for us in a way consistent with other Rails
helpers.

As we go, rename the "old" tag-style helper method to
radio_group_vertical_tag – this is consistent with both the module in
which it appears and the rest of Rails.

It also makes it easy to spot that this approach is deprecated!