Add a form.checkbox_group_vertical helper
Now that checkbox_group_vertical_tag has its new name, that frees us up
to use this method name. Again, delegate to a Rails collection method,
this time collection_check_boxes. Use the mini-builder to get us the div
and class structure we want.

We do need to make sure that the collection we pass to this is fully
stringified key-wise, or comparisons will fail and boxes won't be
checked.