Convert checkbox steps to form_with style
It's a little more convenient to do these at the same time, otherwise
we'd have to make temporary changes to a couple of steps that are reused
in the main feature.

Note that this commit exhibits a quirk of the way we've done things:
We're providing a legend from the view, but we still need to add the
string for it to en.yml. Why? Well, the same string is used for
validation errors.

Thoughts about the best way to deal with this repetition are not fully
formed (translate in the view? Make the helper look it up even though
it's not a label?), so I'm acknowledging that it's happening here.