Let each of our helpers show its own errors
If we can see an error on the object, add a `has-error` class to the
wrapper. Factor this into a private `wrapper_tag` method that looks up
errors (if any), defaulting to a `div` tag but allowing other tags if
necessary. Also responsible for setting the ID, which means that
radio groups and checkbox groups now get an ID (which fixes a bug,
since the error_summary links to errored fields)