Avoid a double label in labelled_text_area
We were relying on the Rails `FormHelper#label` method to look up a
label from translations for us. Unfortunately this was inside another
label. Really we need the translation on its own, which is... well, it's
pretty ugly (nil first param as we don't have an object, need to make
the rest of the params strings, need to point to the idiomatic home for
labels as a scope).