Wrap legends in spans to avoid positioning with float
The default style of the `legend` wants to sit within the top border
of the parent `fieldset`. Floating the element will make sure the
`legend` sits within the `fieldset` however it was ruining the
positioning with the succeeding elements. Wrapping the legend within
a `span` achieves the same outcome without needing a float and keeping
the expected positioning.