Allow tabbing in checkboxes
To restore tab functionality, swap display: none in the
checkbox-group__input for an absolute positioning within its div and an
opacity of 0 such that it responds once again to browser events.

Then style :focus the same as :hover for .checkbox-group__input, and
combine the styles for a selected focussed check box.