Stop 400 errors for blank checkbox lists
By default, Rails outputs a single blank hidden field for collection
check boxes. This is to stop the 400 "parameter missing" error for
strong params. We suppressed it with the include_hidden: false option
to collection_check_boxes because it interfered with the page load
operation of conditional subfields – the "Other" check box was not
showing its associated section.

Reinstate the default hidden field behaviour, and make question_params
strip blank values for methodologies and recording_methods to avoid
validation errors caused by blank values not being valid values for
either.

We will fix the conditional subfield problem in a subsequent commit
(it does not have a failing test so this commit will pass).

While we're here, refactor research_sessions_controller_spec to use the
factory_girl factories (because we can delete a bunch of code)