Make conditional subfields aware of hidden fields
Add a breaking example that represents our Rails use case of a blank
hidden field in a fieldset of checkboxes. Ensure that the controlled
subfield is shown on page load in this circumstance.

Also fix a comment that was out of sync with respect to the behaviour
around clearing field values vs disabling (we're disabling)