Add support for Percy.io
Percy.io adds tha ability to carry out visual regression tests.
This change adds the percy GEM and saves snapshots when going
through the test flow.

Tests run as normal in Circle CI with 2 environment variables set
to tell it which project in Percy.io to save the screenshots to