Move to traits for ResearchSession factories
So far, we've had one solitary set of factories registered with the
same names as all the steps, namely:

* `:researcher`
* `:topic`
* `:purpose` .. and so on, with each factory dependent on the previous.

Now we need an actual `:researcher` factory, that `:researcher`
registered factory name needs to move.

Take the opportunity to move to traits.

## Before

`build_stubbed :researcher` – would return a `ResearchSession` with a
status valid at the `:researcher` step and the researcher details
filled in

## After

`build_stubbed :research_session, :step_researcher` – will return a
`ResearchSession` with a status of `:researcher` and the researcher
details filled in. No longer misleading!