Changelog and rudimentary serving
Comes with no tests at present, given that this was for a little spike.
Add a CHANGELOG.md using the suggested changelog standard for the
vandamme gem (used by Gemnasium). It's not particularly mature, but
it's less important that we agree on its content; more that we agree
on something that communicates our changes.