Remove the concept of age
Since we only talk about the ability to give consent, storing age is
misleading. Instead, use a query string parameter able_to_consent to
determine the preview content at the end of the wizard.

Since this is now a view concern only, move
ResearchSession#able_to_consent? and ResearchSession#unable_to_consent?
to ResearchSessionPresenter. Pass the ability to consent or otherwise
directly from the query string to the presenter in the controller.