Editing from preview returns to preview
When you return to a previous step to edit it from preview, we now pass
an edit-preview=1 parameter on the query string. This in turn passes an
edit-preview=1 value to ResearchSessionsController#update via a hidden
field.

In response, the controller redirects to the preview on save, removing
the need to continue through all the intermediate steps to return to the
preview.