Swapped session date field for simple text
Remove the date type start_datetime attribute and replace it with a
simpler when_text field. Autofocus that field and add a label for it in
en.yml.

Comes with a complicated two-way migration to generate nice-looking text
from the old dates and perform best efforts to roll back should that be
required.

All the work is pretty much making the tests pass.