Visually hide "skip to main content"
When the "skip to main content" link has focus, it moves the content
down the page and appears to the user. We don't want this. Make the
"skip to main content" link remain visually hidden when focused.

Also move the progress bar out of #content so that when the "skip to
main content" link is used, it no longer focuses on the progress bar.