Autofocus the first item on each page
Now we have support for autofocus_first_item in our check box and radio
groups, we can add autofocussed fields to each of our steps.

We do not autofocus the first date time field in time/equipment for two
reasons:

1. It's not supported by the default date control easily
2. We're throwing the date control away pretty much next thing

As such, we'll apply the autofocus to the text box that will replace the
date inputs.

Correct the options hash style for time/equipment while we're here,
though.