Move ResearchSessionController#index to #new
When we visit the root path, we have been serving the page that creates
the form via the #index method. Semantically this has always been a
little dodgy and the addition of a "name" field for a session form shows
this up.

Move it to #new, which is its real semantic home that the code in the
controller action reflects.

Also update #with_stub_routing in the specs – this was working by
accident and these changes break it. We're only interested in
simulating a :create.