Refer to ResearchSessions by #slug everywhere
Put a slugified version of ResearchSession#name in the URL. Ensure we
only ever ask for ResearchSession using find_by(slug: ...)

Ideally we wouldn't be doing this. Ideally we'd be using the form of
routing that relies on to_param and saying
something_path @research_session instead of having to be explicit about
how we refer to ResearchSession everywhere. However, being explicit
would require us to make Question a first-class resource, rather than
a stray :id parameter in a route primarily intended for ResearchSession.