Improve styles for highlights on preview page
Improve the highlighting styles on the preview page to make it less
harsh on the eyes.

The topic / purpose are being wrapped in an additional `p` tag which is
conflicting with the html that is being outputted by `simple_format`. We
should remove the additional `p` tags for now and revise if
`simple_format` is needed.