Make topic/purpose inline/remove formatting
Topic and purpose used to be textareas with too much scope for input.
Now, they're small, focussed fields that have to fit in an inline
sentence structure.

As such, we no longer need formatting of any description, and they are
rendered together in a single paragraph.

Check for the existence of both pieces of content.

Also, use more useful placeholders (not internationalised, as we have no
current idiomatic home for them in the same way as labels)