Define ResearchSession.previewable?
Not really 100% accurate, but good enough for our purposes for now.

Are we on the last step? Then we can preview. This is kind of wrong,
and even as I commit I know we should have a :previewable status that
we set things to after the last step.

La la la la la.