Create a copy
With a really naive '(copy)' disambiguation, create a copy of a session,
but only if at preview stage. Use ResearchSession#previewable? for this,
acknowledging it's *cough* imperfect.

"Create a copy" is just a link to a specialised form of a :new Rails
resource action, where from_existing is passed as a query parameter.

When present, the old session is looked up for the purposes of
suggesting a new name, which it does without any uniqueness checks of
any kind. This "new" form then posts to the create_a_copy action, which
does the grunt work of duplicating the existing session, assigning the
new name (and slug) and redirecting to the preview of the new duplicate.