Add radio buttons to conditionally show optionals
Since we now have DB fields indicating which of the
optional steps are enabled (and hence should have values) use the
standard Barnardos::FormBuilder#radio_group_vertical for each of the
boolean _enabled fields.

Also, take the opportunity to move legends for these radio groups to
en.yml in a new legends area.