Put WhatHappensInThisResearchSession in order
As part of the Great Copy Change (late 2017 edition) allow parts of this
section to remain reactive while using ERB capture to allow us to not
touch methodologies/recording copy yet but still have them in place in
the React component. We expect to be able to remove the calls to
dangerouslySetInnerHTML presently.