Remove anchor & nav styles for the printed sheet
- Remove anchor styles from print styles
- Remove navigation space from styles

When printing the final information sheet and consent form, the
ResearchSession content still has the anchor styles. Participants do not
need to know that this information was linkable before it was printed.
It also disrupts the flow of reading the content.

The navigation is hidden from the final print version however a weird
space is left behind making two green horizontal lines appear.