Apply .eslintrc from digital-transformation repo
We want similar tooling for each of our repos to ensure the developer
experience is consistent. As the digital-transformation repo has the
tooling that a majority of the current team has decided to use, we have
used it as a baseline for how other repos should be.

We should apply the same or similar `.eslintrc` from digital-transformation to the
consent form. We should also fix any linting issues that arise.