Move shared_with_sentences to unvarying_params
To avoid repetition, move shared_with_sentences for the storing step
into the unvarying_params hash in ResearchSessionsHelper. This also
shields us against having to make future changes to SharedWith in three
places.

Because the first of those sentences is consent-sensitive, we need
an extra level of indirection when checking we are able_to_consent?
Only in the situation when we have a ResearchSessionPresenter do we
have access to @research_session.able_to_consent? – in other
circumstances, default to false. This means we will be able to remove
references to able_to_consent from React component parameter lists.