Amend 'able to consent' link to readable content
The 'able to consent' buttons are not a clear solution for a user
choosing which version of the info sheet / consent form they require.
Amend the design so that the user is informed which version they are
currently viewing along with a option to see the other version.

Also wrap element that will not be printed in a new block. This block
should inform the user visually that it is separate to the rest of the
page.